recorder-backend/pyproject.toml
[project]
name = "kielipankki-recorder-backend"
version = "2.0.0"
description = "FastAPI backend for Kielipankki speech donation recorder"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "azure-storage-blob>=12.19.0",
    "aiohttp>=3.9.0",
    "requests>=2.31.0",
    "PyYAML>=6.0",
    "pycryptodome>=3.19.0",
]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "httpx>=0.25.0",  # For testing async endpoints
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["custom_fleep"]
